Add path aliases and move records into domain folder
