+++ src/domain/records/dtos/transfer.dto.ts
@@ -1,7 +1,8 @@
 import { IsNotEmpty, ValidateNested } from 'class-validator';
 import { Type } from 'class-transformer';
-import { CreateExpenseDto } from '../../repositories/expenses/expenses.dto';
-import { CreateIncomeDto } from '../../repositories/incomes/incomes.dto';
+
+import { CreateExpenseDto } from '@repositories/expenses/expenses.dto';
+import { CreateIncomeDto } from '@repositories/incomes/incomes.dto';
 
 export class CreateTransferDto {
   @IsNotEmpty()
