+++ src/domain/records/entities/records.entity.ts
@@ -1,8 +1,8 @@
 import { Prop, Schema, SchemaFactory, raw } from '@nestjs/mongoose';
 import { Document, Types } from 'mongoose';
 
-import { Account as AccountEntity } from '../../repositories/accounts/entities/accounts.entity';
-import { Category } from '../../categories/entities/categories.entity';
+import { Account as AccountEntity } from '@repositories/accounts/entities/accounts.entity';
+import { Category } from '@/categories/entities/categories.entity';
 
 @Schema()
 export class AccountRecord extends Document {
