+++ src/domain/user-accounts-actions/user-actions.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UserAccountsActionsController } from './controller/account-actions.controller';
-import { AccountsModule } from '../../repositories/accounts/accounts.module';
-import { RecordsModule } from '../../records/records.module';
+import { AccountsModule } from '@repositories/accounts/accounts.module';
+import { RecordsModule } from '@domain/records/records.module';
 import { BudgetsModule } from '../../budgets/budgets.module';
 import { BudgetHistoryModule } from '../../budget-history/budget-history.module';
 import { AccountsActionsService } from './services/accounts.-actions.service';
