+++ src/repositories/expenses/expenses.dto.ts
@@ -1,5 +1,5 @@
 import { IsMongoId, IsNotEmpty, IsBoolean, IsArray } from 'class-validator';
-import { CreateRecordDto } from '../../records/dtos/records.dto';
+import { CreateRecordDto } from '@domain/records/dtos/records.dto';
 import { PartialType } from '@nestjs/swagger';
 import { BudgetModel } from '../../budgets/budgets.entity';
