+++ src/repositories/expenses/expenses.entity.ts
@@ -1,6 +1,7 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { AccountRecord } from '../../records/entities/records.entity';
 import { Types } from 'mongoose';
+
+import { AccountRecord } from '@domain/records/entities/records.entity';
 import { BudgetModel } from '../../budgets/budgets.entity';
 
 @Schema()
