+++ src/repositories/incomes/incomes.dto.ts
@@ -1,5 +1,5 @@
 import { IsMongoId, IsNotEmpty, IsArray } from 'class-validator';
-import { CreateRecordDto } from '../../records/dtos/records.dto';
+import { CreateRecordDto } from '@domain/records/dtos/records.dto';
 import { PartialType } from '@nestjs/swagger';
 import { CreateExpense } from '../expenses/expenses.entity';
