+++ src/repositories/incomes/incomes.entity.ts
@@ -1,6 +1,6 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { Types } from 'mongoose';
-import { AccountRecord } from '../../records/entities/records.entity';
+import { AccountRecord } from '@domain/records/entities/records.entity';
 import { CreateExpense } from '../expenses/expenses.entity';
 
 @Schema()
